update and saveField() fix
Updates to make the behavior more Cake 2.x friendly.

Also resolves #3 by checking if Model::data contains all the fields, if
not a Model::read will be performed. A less strict way might be to
check that all aggregate fields are contained instead of the entire
schema, but this is a suitable solution for now.